requirements.txt
flask
flask-mysqldb
pyyaml
Flask-Excel
pyexcel-xlsx
flask_cors
flask_mail

